Module reth_consensus_common::validation

Collection of consensus validation methods. Collection of methods for block validation.

Functions

validate_4844_header_standalone

Validates that the EIP-4844 header fields exist and conform to the spec. This ensures that:

validate_4844_header_with_parent

Validates that the EIP-4844 header fields are correct with respect to the parent block. This ensures that the blob_gas_used and excess_blob_gas fields exist in the child header, and that the excess_blob_gas field matches the expected excess_blob_gas calculated from the parent header fields.

validate_all_transaction_regarding_block_and_nonces

Iterate over all transactions, validate them against each other and against the block. There is no gas check done as REVM already checks that.

validate_block_regarding_chain

Validate block in regards to chain (parent)

validate_block_standalone

Validate a block without regard for state:

validate_header_regarding_parent

Validate block in regards to parent

validate_header_standalone

Validate header standalone

validate_transaction_regarding_header

Validate a transaction in regards to a block header.